feat: use `(U, Const<N>)` as default generic types in `type_array!`
* feat: use `i8` instead of `i32` for exponents

+++ src/name.rs
@@ -32,7 +32,7 @@ impl<N: Name> Debug for N {
     }
 }
 
-pub fn superscript(num: i32) -> String {
+pub fn superscript(num: i8) -> String {
     let s = num.to_string();
     s.bytes()
         .map(|c| match c {
